app/requirements.txt
fastapi
uvicorn[standard]
scikit-learn 
scikit-image 
pandas 
numpy
python-dotenv
jinja2
python-multipart